Adding test to disallow customer usage of v12 storage types
